third_party/WebKit/LayoutTests/http/tests/security/script-crossorigin-onerror-information-expected.txt

CONSOLE MESSAGE: line 1: Uncaught SomeError
The test passes if window.onerror gets unsanitized information about the script error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL msg.match(/SomeError/)[0] should be SomeError. Threw exception TypeError: Cannot read property '0' of null
FAIL url should be http://localhost:8000/security/resources/cors-script.php?fail=true. Was .
FAIL line should be 1. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
